fix: move service clone to internal
